+++ far2l/src/cfg/ConfigOpt.cpp
@@ -147,6 +147,7 @@ const ConfigOpt g_cfg_opts[] {
 	{true,  NSecCmdline, "Splitter", &Opt.CmdLine.Splitter, 1},
 	{true,  NSecCmdline, "WaitKeypress", &Opt.CmdLine.WaitKeypress, 1},
 	{true,  NSecCmdline, "VTLogLimit", &Opt.CmdLine.VTLogLimit, 5000},
+	{false, NSecCmdline, "AskOnMultilinePaste", &Opt.CmdLine.AskOnMultilinePaste, 1},
 
 	{true,  NSecInterface, "Mouse", &Opt.Mouse, 1},
 	{false, NSecInterface, "UseVk_oem_x", &Opt.UseVk_oem_x, 1},
@@ -211,6 +212,7 @@ const ConfigOpt g_cfg_opts[] {
 	{true,  NSecDialog, "MouseButton", &Opt.Dialogs.MouseButton, 0xFFFF},
 	{true,  NSecDialog, "DelRemovesBlocks", &Opt.Dialogs.DelRemovesBlocks, 1},
 	{false, NSecDialog, "CBoxMaxHeight", &Opt.Dialogs.CBoxMaxHeight, 24},
+	{true,  NSecDialog, "ShowArrowsInEdit", &Opt.Dialogs.ShowArrowsInEdit, 1},
 
 	{true,  NSecEditor, "ExternalEditorName", &Opt.strExternalEditor, L""},
 	{true,  NSecEditor, "UseExternalEditor", &Opt.EdOpt.UseExternalEditor, 0},